+++ Timer/MVUserDefaultsKeys.swift
@@ -1,4 +1,4 @@
-struct MVUserDefaultsKeys {
+enum MVUserDefaultsKeys {
   static let staysOnTop = "staysOnTop"
   static let soundIndex = "soundIndex"
 }